+++ src/main/java/com/org/bgv/entity/RuleTypes.java
@@ -34,7 +34,7 @@ public class RuleTypes {
 	 private String name;
 	 private String code;
 	 private String label;
-	 private Long minCount;
-	 private Long maxCount;
+	 private Integer minCount;
+	 private Integer maxCount;
 	
 }
